backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.1
black==24.10.0
boto3==1.34.54
botocore==1.34.143
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
exceptiongroup==1.2.1
fastapi==0.110.0
grpcio==1.65.0
grpcio-health-checking==1.65.0
grpcio-tools==1.65.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
loguru==0.7.2
mypy-extensions==1.0.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
protobuf==5.27.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.27.1
validators==0.33.0
weaviate-client==4.7.1
